requirements.txt
flask
mysql-connector-python
click
mysql-db